Time: 22 ms (10.04%), Space: 22.9 MB (19.84%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -15,7 +15,7 @@ public:
 
             int kcan = 0;
             for (int num: piles){
-                kcan += ceil((double)num/(double)mid);
+                kcan += ceil((double)num/mid);
             }
 
             if (kcan <= h){
